requirements.txt
flask==3.0.3
groq
python-dotenv
gunicorn
